enter to comment, change rating

+++ db/migrate/20190717073237_create_rates.rb
@@ -1,7 +1,7 @@
 class CreateRates < ActiveRecord::Migration
   def change
     create_table :rates do |t|
-      t.integer :score
+      t.float :score
       t.references :photo, index: true, foreign_key: true
       t.references :user, index: true, foreign_key: true
